refactor(server): decompose AndroidGattServerCallback (444→126 loc)
Split the 444-line God File into focused handler files using
the Class Decomposition Pattern (Approach B: Extension Functions).

- AndroidGattServerCallback.kt: thin callback wrapper (126 loc)
- AndroidGattServerConnectionHandlers.kt: connection + MTU (46 loc)
- AndroidGattServerReadHandlers.kt: read, descriptor read, prepared writes (161 loc)
- AndroidGattServerWriteHandlers.kt: write, descriptor write, execute, CCCD (227 loc)

All files <300 lines. state and sendResponseSafe made internal for
same-package extension function access.

Closes #241-P2
